@@ -1,71 +1,72 @@
 {
-	"name": "finances-easy-web",
-	"version": "1.0.0",
-	"license": "Apache",
-	"scripts": {
-		"ng": "ng",
-		"start": "node server.js",
-		"build": "ng build --prod",
-		"gitbuild": "ng build --prod --base-href /dist/",
-		"test": "ng test",
-		"test-ci": "TEST_CI=true ng test",
-		"lint": "ng lint",
-		"e2e": "ng e2e",
-		"heroku-postbuild": "ng build --prod"
-	},
-	"private": true,
-	"dependencies": {
-		"@angular/animations": "7.0.2",
-		"@angular/cli": "^7.2.3",
-		"@angular/common": "7.0.2",
-		"@angular/compiler": "7.0.2",
-		"@angular/compiler-cli": "^7.2.6",
-		"@angular/core": "7.0.2",
-		"@angular/forms": "7.0.2",
-		"@angular/http": "7.0.2",
-		"@angular/platform-browser": "7.0.2",
-		"@angular/platform-browser-dynamic": "7.0.2",
-		"@angular/router": "7.0.2",
-		"@ng-bootstrap/ng-bootstrap": "4.0.0",
-		"@ngrx/core": "^1.2.0",
-		"@ngrx/effects": "^7.2.0",
-		"@ngrx/router-store": "^7.2.0",
-		"@ngrx/store": "^7.2.0",
-		"@ngrx/store-devtools": "^7.2.0",
-		"@ngx-translate/core": "11.0.0",
-		"@ngx-translate/http-loader": "4.0.0",
-		"angular-calendar": "^0.26.11",
-		"chart.js": "2.7.3",
-		"core-js": "2.5.7",
-		"date-fns": "^1.30.1",
-		"express": "^4.16.4",
-		"font-awesome": "4.7.0",
-		"moment": "^2.24.0",
-		"ng2-charts": "2.2.3",
-		"ngx-ui-loader": "^7.2.1",
-		"rxjs": "6.3.3",
-		"yarn": "1.12.1",
-		"zone.js": "0.8.26"
-	},
-	"devDependencies": {
-		"@angular-devkit/build-angular": "^0.13.9",
-		"@angular/language-service": "7.0.2",
-		"@types/jasmine": "2.8.9",
-		"@types/jasminewd2": "2.0.5",
-		"@types/node": "10.12.2",
-		"codelyzer": "4.5.0",
-		"jasmine-core": "3.3.0",
-		"jasmine-spec-reporter": "4.2.1",
-		"karma": "^4.0.1",
-		"karma-chrome-launcher": "2.2.0",
-		"karma-cli": "1.0.1",
-		"karma-coverage-istanbul-reporter": "2.0.4",
-		"karma-jasmine": "1.1.2",
-		"karma-jasmine-html-reporter": "1.4.0",
-		"protractor": "5.4.1",
-		"ts-node": "7.0.1",
-		"tslint": "5.11.0",
-		"typescript": "^3.2.4"
-	},
-	"heroku-run-build-script": true
+  "name": "finances-easy-web",
+  "version": "1.0.0",
+  "license": "Apache",
+  "scripts": {
+    "ng": "ng",
+    "start": "node server.js",
+    "build": "ng build --prod",
+    "gitbuild": "ng build --prod --base-href /dist/",
+    "test": "ng test",
+    "test-ci": "TEST_CI=true ng test",
+    "lint": "ng lint",
+    "e2e": "ng e2e",
+    "heroku-postbuild": "ng build --prod"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "8.0.0",
+    "@angular/cli": "^7.3.9",
+    "@angular/common": "8.0.0",
+    "@angular/compiler": "8.0.0",
+    "@angular/compiler-cli": "^8.0.0",
+    "@angular/core": "8.0.0",
+    "@angular/forms": "8.0.0",
+    "@angular/http": "7.0.2",
+    "@angular/platform-browser": "8.0.0",
+    "@angular/platform-browser-dynamic": "8.0.0",
+    "@angular/router": "8.0.0",
+    "@ng-bootstrap/ng-bootstrap": "4.0.0",
+    "@ngrx/core": "^1.2.0",
+    "@ngrx/effects": "^7.4.0",
+    "@ngrx/router-store": "^7.4.0",
+    "@ngrx/store": "^7.4.0",
+    "@ngrx/store-devtools": "^7.4.0",
+    "@ngx-translate/core": "11.0.0",
+    "@ngx-translate/http-loader": "4.0.0",
+    "angular-calendar": "^0.26.11",
+    "chart.js": "2.7.3",
+    "core-js": "2.5.7",
+    "date-fns": "^1.30.1",
+    "express": "^4.17.1",
+    "font-awesome": "4.7.0",
+    "moment": "^2.24.0",
+    "ng2-charts": "2.2.3",
+    "ngx-ui-loader": "^7.2.2",
+    "rxjs": "^6.5.2",
+    "tslib": "^1.9.0",
+    "yarn": "1.12.1",
+    "zone.js": "0.9.1"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "~0.13.0",
+    "@angular/language-service": "8.0.0",
+    "@types/jasmine": "2.8.9",
+    "@types/jasminewd2": "2.0.5",
+    "@types/node": "10.12.2",
+    "codelyzer": "4.5.0",
+    "jasmine-core": "3.3.0",
+    "jasmine-spec-reporter": "4.2.1",
+    "karma": "^4.1.0",
+    "karma-chrome-launcher": "2.2.0",
+    "karma-cli": "1.0.1",
+    "karma-coverage-istanbul-reporter": "2.0.4",
+    "karma-jasmine": "1.1.2",
+    "karma-jasmine-html-reporter": "1.4.0",
+    "protractor": "5.4.1",
+    "ts-node": "7.0.1",
+    "tslint": "5.11.0",
+    "typescript": "^3.4.5"
+  },
+  "heroku-run-build-script": true
 }
