@@ -2,12 +2,13 @@ import { HttpClient, HttpParams } from '@angular/common/http';
 import { Injectable } from '@angular/core';
 import { Actions, Effect, ofType } from '@ngrx/effects';
 import { of, pipe } from 'rxjs';
-import { catchError, map, switchMap } from 'rxjs/operators';
+import { catchError, map, mergeMap, switchMap } from 'rxjs/operators';
 import { environment } from '../../../../environments/environment';
 import { Default } from '../../../shared/enum/default.enum';
 import { Expense } from '../../../shared/model/expense.model';
 import { AlertActionsEnum, ShowAlertError } from '../../../store/alert.actions';
 import * as ExpenseActions from './expense.actions';
+import { ExpenseActionsEnum } from './expense.actions';
 
 
 
@@ -20,7 +21,7 @@ export class ExpenseEffects {
 
 	@Effect()
 	createExpense = this.actions.pipe(
-		ofType(ExpenseActions.ExpenseActionsEnum.CREATE_EXPENSE),
+		ofType(ExpenseActionsEnum.CREATE_EXPENSE),
 		map((action: ExpenseActions.CreateExpense) => {
 			return action.payload;
 		}),
@@ -28,11 +29,17 @@ export class ExpenseEffects {
 			switchMap((expense: Expense) => {
 				return this.http.post<Expense>(this.expenseEndPoint.concat('/create'), expense);
 			}),
-			map(() => {
-				return {
-					type: AlertActionsEnum.SHOW_ALERT_SUCESS,
-					payload: 'Expense saved'
-				};
+			mergeMap((expense: Expense) => {
+				return [
+					{
+						type: ExpenseActionsEnum.ADD_EXPENSE,
+						payload: expense
+					},
+					{
+						type: AlertActionsEnum.SHOW_ALERT_SUCESS,
+						payload: 'Expense saved'
+					}
+				];
 			}),
 			catchError(() => {
 				return of(new ShowAlertError('Error on save expense'));
@@ -47,14 +54,20 @@ export class ExpenseEffects {
 			return action.payload;
 		}),
 		pipe(
-			switchMap((inc: Expense) => {
-				return this.http.post<Expense>(this.expenseEndPoint.concat('/update'), inc);
+			switchMap((exp: Expense) => {
+				return this.http.post<Expense>(this.expenseEndPoint.concat('/update'), exp);
 			}),
-			map(() => {
-				return {
-					type: AlertActionsEnum.SHOW_ALERT_SUCESS,
-					payload: 'Expense edited'
-				};
+			mergeMap((exp: Expense) => {
+				return [
+					{
+						type: ExpenseActionsEnum.ALTER_EXPENSE,
+						payload: exp
+					},
+					{
+						type: AlertActionsEnum.SHOW_ALERT_SUCESS,
+						payload: 'Expense edited'
+					}
+				];
 			}),
 			catchError(() => {
 				return of(new ShowAlertError('Error on edit expense'));
@@ -72,11 +85,17 @@ export class ExpenseEffects {
 			switchMap((id: number) => {
 				return this.http.delete(this.expenseEndPoint.concat('/delete/').concat(id.toString()));
 			}),
-			map(() => {
-				return {
-					type: AlertActionsEnum.SHOW_ALERT_SUCESS,
-					payload: 'Expense deleted'
-				};
+			mergeMap((id: number) => {
+				return [
+					{
+						type: ExpenseActionsEnum.REMOVE_EXPENSE,
+						payload: id
+					},
+					{
+						type: AlertActionsEnum.SHOW_ALERT_SUCESS,
+						payload: 'Expense deleted'
+					}
+				];
 			}),
 			catchError(() => {
 				return of(new ShowAlertError('Error on delete expense'));
